refactor($compile): 문자열 포장으로 리팩토링
- carName을 포장하는 class 생성
- 5자 validation CarName클래스로 이관
- 테스트 케이스 수정
